requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.0.1
billiard==4.2.4
black==25.1.0
celery==5.6.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
DateTime==6.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.128.0
fastapi-cache2==0.2.2
greenlet==3.2.2
h11==0.16.0
idna==3.10
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.2.0
pillow==12.1.1
platformdirs==4.3.8
prompt_toolkit==3.0.52
pydantic==2.11.2
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.22
pytz==2025.2
redis==7.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.50.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.2
